llm-flask-rag-azure/requirements.txt
langchain==0.1.11
langchain-openai==0.0.5
langchain-community==0.0.27
python-dotenv==1.0.1
flask==3.0.2
openai==1.11.1
pypdf==4.0.1
sentence_transformers==2.2.2
qdrant-client==1.8.2
redis==5.0.3
